[SourceKit/CodeFormat] Indent lines in multi-line strings
The indentation of a new line in a multi-line string will now be the
first previous line that had contents, except for when:
  1. that line is the first line in the string and that line isn't only
     quotes, or
  2. the resuling indentation would be invalid (ie. less than the end
     quotes)

For (1), an extra indentation level to the start line will be used as
long as there are no end quotes. Otherwise if there are, or for (2), the
indentation will be the same as that of the end quotes.

Lines that have content (whether in an existing string or not) will not
be formatted at all, unless they're currently invalid (ie. their
indentation level is less than that of their end quotes).

Resolves rdar://32181422