Remove use of getClangDecl in Refactoring
`getClangDecl` was being used to check whether or not a given decl was a
reference to an imported ObjC decl so that renaming could be skipped.

Use `originatedFromClang` instead. This also handles the error wrapper
struct that doesn't actually have an underlying Clang node, but should
not be able to be renamed.